move BrowserESModuleLoader to single file
